+++ src/main/java/net/meander/subtlyd/util/data/tags/DamageTypeTagsSD.java
@@ -27,7 +27,7 @@ public class DamageTypeTagsSD extends KeyTagProvider<DamageType> {
         this.tag(ALWAYS_KILLS_TENT).add(DamageTypes.ARROW, DamageTypes.TRIDENT, DamageTypes.FIREBALL, DamageTypes.WITHER_SKULL, DamageTypes.WIND_CHARGE);
         this.tag(BURNS_TENTS).add(DamageTypes.ON_FIRE);
         this.tag(IGNITES_TENTS).add(DamageTypes.IN_FIRE, DamageTypes.CAMPFIRE);
-        this.tag(IS_OCCULT).add(DamageTypes.MAGIC, DamageTypes.INDIRECT_MAGIC, DamageTypes.SONIC_BOOM, DamageTypes.THORNS, DamageTypes.WITHER);
+        this.tag(IS_OCCULT).add(DamageTypes.MAGIC, DamageTypes.INDIRECT_MAGIC, DamageTypes.SONIC_BOOM, DamageTypes.THORNS);
     }
 
     private static TagKey<DamageType> create(String string) {