+++ src/main/java/net/meander/subtlyd/world/item/enchantment/EnchantmentsSD.java
@@ -3,6 +3,7 @@ package net.meander.subtlyd.world.item.enchantment;
 import net.meander.subtlyd.util.Util;
 import net.meander.subtlyd.util.data.tags.DamageTypeTagsSD;
 import net.minecraft.advancements.criterion.DamageSourcePredicate;
+import net.minecraft.advancements.criterion.ItemPredicate;
 import net.minecraft.advancements.criterion.TagPredicate;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
@@ -18,9 +19,12 @@ import net.minecraft.world.item.enchantment.EnchantmentEffectComponents;
 import net.minecraft.world.item.enchantment.LevelBasedValue;
 import net.minecraft.world.item.enchantment.effects.AddValue;
 import net.minecraft.world.level.storage.loot.predicates.DamageSourceCondition;
+import net.minecraft.world.level.storage.loot.predicates.InvertedLootItemCondition;
+import net.minecraft.world.level.storage.loot.predicates.MatchTool;
 
 public class EnchantmentsSD {
     public static final ResourceKey<Enchantment> OCCULT_PROTECTION = ResourceKey.create(Registries.ENCHANTMENT, Util.identifier("occult_protection"));
+    public static final ResourceKey<Enchantment> DECAYING_CURSE = ResourceKey.create(Registries.ENCHANTMENT, Util.identifier("decaying_curse"));
 
     public static void bootstrap(BootstrapContext<Enchantment> context) {
         HolderGetter<Item> items = context.lookup(Registries.ITEM);
@@ -39,5 +43,19 @@ public class EnchantmentsSD {
                                 DamageSourceCondition.hasDamageSource(DamageSourcePredicate.Builder.damageType().tag(TagPredicate.is(DamageTypeTagsSD.IS_OCCULT))
                                         .tag(TagPredicate.isNot(DamageTypeTags.BYPASSES_INVULNERABILITY))))
                 .build(OCCULT_PROTECTION.identifier()));
+        context.register(DECAYING_CURSE, Enchantment.enchantment(
+                                Enchantment.definition(
+                                        items.getOrThrow(ItemTags.DURABILITY_ENCHANTABLE),
+                                        1,
+                                        1,
+                                        Enchantment.constantCost(25),
+                                        Enchantment.constantCost(50),
+                                        8,
+                                        EquipmentSlotGroup.ANY))
+                        .withEffect(EnchantmentEffectComponents.ITEM_DAMAGE, new AddValue(LevelBasedValue.constant(1.0F)),
+                                MatchTool.toolMatches(ItemPredicate.Builder.item().of(items, ItemTags.ARMOR_ENCHANTABLE)))
+                        .withEffect(EnchantmentEffectComponents.ITEM_DAMAGE, new AddValue(LevelBasedValue.constant(1.0F)),
+                                InvertedLootItemCondition.invert(MatchTool.toolMatches(ItemPredicate.Builder.item().of(items, ItemTags.ARMOR_ENCHANTABLE))))
+                .build(DECAYING_CURSE.identifier()));
     }
 }